requirements.txt
django==5.0.8
djangorestframework==3.15.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
django-jazzmin==3.0.0
pydantic==2.9.2
bcrypt==4.2.0
PyJWT==2.9.0
email-validator==2.2.0